@@ -5,9 +5,10 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
+    "builddev": "vue-cli-service build --mode development",
+    "test:unit": "jest --clearCache && vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
-    "lint:pug": "pug-lint-vue src",
-    "test:unit": "jest --clearCache && vue-cli-service test:unit"
+    "lint:pug": "pug-lint-vue src"
   },
   "dependencies": {
     "@girder/components": "v3.0.0-alpha.3",
@@ -15,8 +16,10 @@
     "@openfonts/barlow-condensed_all": "^0.1.1",
     "@sentry/browser": "^5.0.7",
     "@sentry/integrations": "^5.0.7",
+    "@vue/composition-api": "^0.6.4",
     "axios": "^0.18.0",
     "c3": "^0.7.8",
+    "core-js": "^3.6.5",
     "d3-array": "^2.3.1",
     "d3-axis": "^1.0.12",
     "d3-boxplot": "^0.7.5",
@@ -47,25 +50,36 @@
     "vuex": "^3.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.4",
-    "@vue/cli-plugin-eslint": "^3.0.4",
-    "@vue/cli-plugin-unit-jest": "^3.5.3",
-    "@vue/cli-service": "^3.0.4",
-    "@vue/eslint-config-airbnb": "^4.0.0",
+    "@types/d3": "^5.7.2",
+    "@types/d3-scale": "^2.2.0",
+    "@types/d3-selection": "^1.4.1",
+    "@types/jest": "^24.0.19",
+    "@types/lodash": "^4.14.157",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
+    "@vue/cli-plugin-babel": "~4.4.4",
+    "@vue/cli-plugin-eslint": "~4.4.4",
+    "@vue/cli-plugin-typescript": "~4.4.0",
+    "@vue/cli-plugin-unit-jest": "~4.4.4",
+    "@vue/cli-service": "~4.4.4",
+    "@vue/eslint-config-airbnb": "^5.0.2",
+    "@vue/eslint-config-typescript": "^5.0.2",
     "@vue/test-utils": "^1.0.0-beta.29",
     "babel-core": "~7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^24.7.0",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-vue": "^6.2.2",
     "git-revision-webpack-plugin": "^3.0.3",
     "node-sass": "^4.13.0",
     "pug": "^2.0.3",
     "pug-lint-vue": "^0.1.3",
     "pug-plain-loader": "^1.0.0",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^8.0.2",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
+    "typescript": "~3.9.3",
     "vue-template-compiler": "^2.5.21",
     "webpack": "^4.29.6"
   },
@@ -75,8 +89,10 @@
       "node": true
     },
     "extends": [
+      "eslint:recommended",
       "plugin:vue/recommended",
-      "@vue/airbnb"
+      "@vue/airbnb",
+      "@vue/typescript/recommended"
     ],
     "rules": {
       "camelcase": 0,
@@ -87,15 +103,29 @@
         }
       ],
       "no-underscore-dangle": 0,
-      "import/prefer-default-export": 0
+      "import/prefer-default-export": 0,
+      "@typescript-eslint/camelcase": 0,
+      "@typescript-eslint/no-empty-function": 0,
+      "@typescript-eslint/ban-ts-ignore": 0
     },
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@typescript-eslint/parser",
+      "ecmaVersion": 2020
     },
     "globals": {
       "COMMITHASH": true
     }
   },
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {}
+    }
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not ie <= 8"
+  ],
   "jest": {
     "cache": false,
     "moduleFileExtensions": [
@@ -116,7 +146,8 @@
     ],
     "testMatch": [
       "**/tests/**/*.spec.js"
-    ]
+    ],
+    "preset": "@vue/cli-plugin-unit-jest/presets/typescript-and-babel"
   },
   "pugLintConfig": {
     "disallowAttributeConcatenation": true,
@@ -150,15 +181,5 @@
     "validateLineBreaks": "LF",
     "validateSelfClosingTags": true,
     "validateTemplateString": true
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ]
+  }
 }
