@@ -1,9 +1,12 @@
+from typing import Optional
+
 import numpy as np
+import pandas as pd
 
 SCALING_METHODS = {'auto', 'range', 'pareto', 'vast', 'level'}
 
 
-def scale(method, table):
+def scale(method: Optional[str], table: pd.DataFrame) -> pd.DataFrame:
     if method is None:
         pass
     elif method == 'auto':
@@ -21,28 +24,28 @@ def scale(method, table):
     return table
 
 
-def center(table):
+def center(table: pd.DataFrame) -> pd.DataFrame:
     return table - table.mean()
 
 
-def auto(table):
+def auto(table: pd.DataFrame) -> pd.DataFrame:
     return center(table) / table.std()
 
 
-def range(table):
+def range(table: pd.DataFrame) -> pd.DataFrame:
     return center(table) / (table.max() - table.min())
 
 
-def pareto(table):
+def pareto(table: pd.DataFrame) -> pd.DataFrame:
     return center(table) / np.sqrt(table.std())
 
 
-def vast(table):
+def vast(table: pd.DataFrame) -> pd.DataFrame:
     mean = table.mean()
     sd = table.std()
     return (mean / sd) * (table - mean) / sd
 
 
-def level(table):
+def level(table: pd.DataFrame) -> pd.DataFrame:
     mean = table.mean()
     return (table - mean) / mean
