@@ -1,9 +1,13 @@
+from typing import Optional
+
 import numpy as np
+import pandas as pd
 
 TRANSFORMATION_METHODS = {'log10', 'squareroot', 'cuberoot', 'log2'}
 
 
-def transform(method, table):
+def transform(method: Optional[str], table: pd.DataFrame) -> pd.DataFrame:
+    table = table.astype(np.float64)
     if method is None:
         table = table
     elif method == 'log10':
@@ -19,17 +23,17 @@ def transform(method, table):
     return table
 
 
-def log10(table, min=1e-8):
-    return np.log(table.clip(lower=min)) / np.log(10)
+def log10(table: pd.DataFrame, min=1e-8):
+    return np.log10(table.clip(lower=min))
 
 
-def log2(table, min=1e-8):
-    return np.log(table.clip(lower=min)) / np.log(2)
+def log2(table: pd.DataFrame, min=1e-8):
+    return np.log2(table.clip(lower=min))
 
 
-def squareroot(table):
+def squareroot(table: pd.DataFrame):
     return np.sqrt(table.clip(lower=0))
 
 
-def cuberoot(table):
+def cuberoot(table: pd.DataFrame):
     return np.power(table.clip(lower=0), 1.0 / 3.0)
