@@ -1,12 +1,13 @@
-
-<template lang="pug">
-router-view
+<template>
+  <router-view />
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import Vue from 'vue';
+
+export default Vue.extend({
   name: 'App',
-};
+});
 </script>
 
 <style lang="scss">
