@@ -3,9 +3,9 @@ import axios from 'axios';
 import VueAxios from 'vue-axios';
 
 const ApiService = {
-  init() {
+  init(serverURL) {
     Vue.use(VueAxios, axios);
-    Vue.axios.defaults.baseURL = '/api/v1';
+    Vue.axios.defaults.baseURL = serverURL;
   },
 
   buildUrl(path, args) {
@@ -126,7 +126,6 @@ export const CSVService = {
   },
 };
 
-
 export const ExcelService = {
   upload(file) {
     const formData = new FormData();
@@ -134,7 +133,6 @@ export const ExcelService = {
     return ApiService.upload('excel', formData);
   },
 };
-
 
 export const SampleService = {
 
