@@ -1,67 +1,90 @@
-<script>
+<script lang="ts">
+import {
+  defineComponent, toRefs, PropType, computed,
+} from '@vue/composition-api';
 import ColorerOption from './ColorerOption.vue';
-import sampleMixin from './mixins/sampleMixin';
+import useSample, { sampleProps } from './use/useSample';
 
-export default {
-  components: {
-    ColorerOption,
-  },
-  mixins: [sampleMixin],
+interface Colorer {
+  option: string | null;
+  levels: {
+    name: string;
+    color: string[];
+  };
+  apply?: (row: string) => string | null;
+}
+
+export default defineComponent({
   props: {
     title: {
       type: String,
       required: false,
       default: 'Sample Color',
     },
-    value: { // {option: string | null,
-    // levels: {name: string, color: string}[], apply(row: string) => string | null}
-      type: Object,
+    value: {
+      type: Object as PropType<Colorer>,
       required: false,
       default: null,
     },
+    ...sampleProps,
   },
-  computed: {
-    validatedValue() {
-      if (this.value) {
-        return this.value.option;
-      }
-      if (this.options.length === 0) {
-        return null;
-      }
-      const v = this.options[0].value;
-      this.changeValue(v);
-      return v;
-    },
+  components: {
+    ColorerOption,
   },
-  methods: {
-    generateColorer(value) {
+  setup(props, context) {
+    const { options, categoricalMetaData, rowToIndex } = useSample(toRefs(props), context);
+    function generateColorer(value: string) {
       if (!value) {
-        return () => null;
+        return () => undefined;
       }
-      const meta = this.categoricalMetaData.find(d => d.value === value);
+      const meta = categoricalMetaData.value.find((d) => d.value === value);
+      if (!meta) {
+        return () => undefined;
+      }
       const lookup = new Map(meta.levels.map(({ name, color }) => [name, color]));
-      const toIndex = this.rowToIndex;
-      return column => lookup.get(meta.data[toIndex(column)]);
-    },
-    generateLevels(value) {
-      if (!value || value === this.emptyOption) {
+      const toIndex = rowToIndex.value;
+      return (column: string) => lookup.get(meta.data[toIndex(column)]);
+    }
+    function generateLevels(value: string) {
+      if (!value || value === props.emptyOption) {
         return [];
       }
-      return this.options.find(d => d.value === value).options;
-    },
-    changeValue(value) {
+      return options.value.find((d) => d.value === value)?.options || [];
+    }
+    function changeValue(value: string) {
       const wrapper = {
         option: value,
-        levels: this.generateLevels(value),
-        apply: this.generateColorer(value),
+        levels: generateLevels(value),
+        apply: generateColorer(value),
       };
-      this.$emit('input', wrapper);
-    },
+      context.emit('input', wrapper);
+    }
+    const validatedValue = computed(() => {
+      if (props.value) {
+        return props.value.option;
+      }
+      if (options.value.length === 0) {
+        return null;
+      }
+      const v = options.value[0].value;
+      changeValue(v);
+      return v;
+    });
+    return {
+      options,
+      validatedValue,
+      changeValue,
+    };
   },
-};
+});
 </script>
 
-<template lang="pug">
-colorer-option(:title="title", :disabled="disabled",
-    :options="options", :value="validatedValue", @input="changeValue($event)")
+<template>
+  <colorer-option
+    :title="title"
+    :disabled="disabled"
+    :options="options"
+    :value="validatedValue"
+    @input="changeValue($event)"
+  />
 </template>
