@@ -36,11 +36,11 @@ export default {
       }
       let data = df.columnNames.map((name, i) => ({
         name,
-        values: df.data.map(row => row[i]),
+        values: df.data.map((row) => row[i]),
       }));
 
       if (this.metaboliteFilter && this.metaboliteFilter.option) {
-        data = data.filter(d => this.metaboliteFilter.apply(d.name));
+        data = data.filter((d) => this.metaboliteFilter.apply(d.name));
       }
 
       if (this.sampleFilter && this.sampleFilter.option) {
@@ -49,10 +49,10 @@ export default {
         data.forEach((row) => {
           const vs = row.values;
           delete row.values;
-          row.groups = groups.map(group => ({
+          row.groups = groups.map((group) => ({
             name: group.name,
             color: group.color,
-            values: group.indices.map(i => vs[i]),
+            values: group.indices.map((i) => vs[i]),
           }));
         });
       }
@@ -61,7 +61,7 @@ export default {
     },
 
     groups() {
-      return this.sampleFilter && this.sampleFilter.option ? this.sampleFilter.filter : null;
+      return this.sampleFilter && this.sampleFilter.option ? this.sampleFilter.filter : [];
     },
   },
 };
@@ -69,8 +69,13 @@ export default {
 </script>
 
 <template lang="pug">
-vis-tile-large(v-if="dataset", title="Metabolite Box Plot", :loading="false",
-    download, expanded)
+vis-tile-large(
+    v-if="dataset",
+    title="Metabolite Box Plot",
+    analysis-path="boxplot",
+    :loading="false",
+    download,
+    expanded)
   template(#controls)
     metabolite-filter(:dataset="dataset", v-model="metaboliteFilter")
     sample-filter(:dataset="dataset", v-model="sampleFilter",
@@ -78,7 +83,8 @@ vis-tile-large(v-if="dataset", title="Metabolite Box Plot", :loading="false",
 
   boxplot-plot.main(
       v-if="dataset.ready",
-      :rows="chartData", :groups="groups")
+      :rows="chartData",
+      :groups="groups")
 </template>
 
 <style scoped>
