@@ -10,7 +10,7 @@ import { svg2url } from '../../utils/exporter';
 function domain(arr) {
   // let min = Number.POSITIVE_INFINITY;
   let max = Number.NEGATIVE_INFINITY;
-  arr.forEach(row => row.forEach((v) => {
+  arr.forEach((row) => row.forEach((v) => {
     // if (v < min) {
     //   min = v;
     // }
@@ -44,15 +44,21 @@ function aggregate(arr, rs, cs, transposed) {
   return sum / l;
 }
 
-const DENDOGRAM_RATIO = 0.2;
+const DENDROGRAM_RATIO = 0.2;
 const LABEL_WIDTH = 150;
 
 const MDI_PLUS_CIRCLE = '\uF417;';
 const MDI_MINUS_CIRCLE = '\uF376;';
 const MDI_STAR_CIRCLE = '\uF4CF;';
 
+// max font size in px for metabolite names.
+// Anything larger will cause metabolite
+// names to be misaligned
+const MAX_FONT_SIZE = 12;
+
 export const heatmapLayouts = [
   { label: 'Auto', value: 'auto' },
+  { label: 'Full', value: 'full' },
   { label: 'Square Cells', value: 'squareCells' },
   { label: 'Square Matrix', value: 'squareMatrix' },
 ];
@@ -81,7 +87,7 @@ function changeHovered(wrapper, arr) {
     wrapper.hovered = new Set(arr);
     return true;
   }
-  if (old.size === 0 || ((arr.length === 1 && old.has(arr[0])) || arr.every(d => old.has(d)))) {
+  if (old.size === 0 || ((arr.length === 1 && old.has(arr[0])) || arr.every((d) => old.has(d)))) {
     return false;
   }
 
@@ -98,6 +104,10 @@ export default {
       type: Array,
       default: () => [],
     },
+    original: { // { columns: string[], rows: string[], values: number[][] }
+      type: Object,
+      default: () => {},
+    },
     columnClustering: { // ITreeNode
       type: Object,
       default: null,
@@ -106,17 +116,17 @@ export default {
       type: Object,
       default: null,
     },
-    rowConfig: { // { dendogram: boolean, colorer?: (name) => string }
+    rowConfig: { // { dendrogram: boolean, colorer?: (name) => string }
       type: Object,
-      default: () => ({ dendogram: true, colorer: null }),
+      default: () => ({ dendrogram: true, colorer: null }),
     },
-    columnConfig: { // { dendogram: boolean, colorer?: (name) => string }
+    columnConfig: { // { dendrogram: boolean, colorer?: (name) => string }
       type: Object,
-      default: () => ({ dendogram: true, colorer: null }),
+      default: () => ({ dendrogram: true, colorer: null }),
     },
-    layout: { // { dendogram: boolean }
+    layout: { // { dendrogram: boolean }
       type: String,
-      validate: v => heatmapLayouts.find(d => d.value === v),
+      validate: (v) => heatmapLayouts.find((d) => d.value === v),
       default: heatmapLayouts[0].value,
     },
     transposed: {
@@ -143,7 +153,7 @@ export default {
         collapsed: new Set(),
         focus: null,
       },
-      DENDOGRAM_RATIO,
+      DENDROGRAM_RATIO,
       LABEL_WIDTH,
     };
   },
@@ -212,32 +222,82 @@ export default {
       return root;
     },
     valueScale() {
-      return scaleSequential(t => interpolateRdBu(1 - t)).domain(domain(this.values));
+      return scaleSequential((t) => interpolateRdBu(1 - t)).domain(domain(this.values));
     },
     legendDomain() {
       return this.valueScale.domain().map(this.format);
     },
+
     columnLeaves() {
+      return this.columnConfig.dendrogram ? this.columnLeavesSorted : this.columnLeavesOriginal;
+    },
+
+    columnLeavesSorted() {
       return this.columnTree ? this.columnTree.leaves() : [];
     },
+
+    columnLeavesOriginal() {
+      const {
+        columnLeavesSorted,
+        original,
+        transposed,
+      } = this;
+
+      // Make a lookup table of the leaves by leaf name.
+      const lookup = new Map();
+      columnLeavesSorted.forEach((leaf) => {
+        lookup.set(leaf.data.name, leaf);
+      });
+
+      // Reorder the leaves by the order given in the clustering prop.
+      const source = transposed ? original.rows : original.columns;
+      return source.map((name) => lookup.get(name));
+    },
+
     rowLeaves() {
+      return this.rowConfig.dendrogram ? this.rowLeavesSorted : this.rowLeavesOriginal;
+    },
+
+    rowLeavesSorted() {
       return this.rowTree ? this.rowTree.leaves() : [];
     },
-    columnDendogramHeight() {
-      return this.columnConfig.dendogram ? this.height * DENDOGRAM_RATIO : 0;
+
+    rowLeavesOriginal() {
+      const {
+        rowLeavesSorted,
+        original,
+        transposed,
+      } = this;
+
+      // Make a lookup table of the leaves by leaf name.
+      const lookup = new Map();
+      rowLeavesSorted.forEach((leaf) => {
+        lookup.set(leaf.data.name, leaf);
+      });
+
+      // Reorder the leaves by the order given in the clustering prop.
+      const source = transposed ? original.columns : original.rows;
+      return source.map((name) => lookup.get(name));
     },
-    rowDendogramWidth() {
-      return this.rowConfig.dendogram ? this.width * DENDOGRAM_RATIO : 0;
+
+    columnDendrogramHeight() {
+      return this.columnConfig.dendrogram ? this.height * DENDROGRAM_RATIO : 0;
+    },
+    rowDendrogramWidth() {
+      return this.rowConfig.dendrogram ? this.width * DENDROGRAM_RATIO : 0;
     },
     matrixDimensions() {
-      let width = this.width - this.rowDendogramWidth - LABEL_WIDTH;
-      let height = this.height - this.columnDendogramHeight - LABEL_WIDTH;
+      let width = Math.max(this.width - this.rowDendrogramWidth - LABEL_WIDTH, 0);
+      let height = Math.max(this.height - this.columnDendrogramHeight - LABEL_WIDTH, 0);
       if (this.layout === 'squareCells') {
         const wx = width / this.columnLeaves.length;
         const hy = height / this.rowLeaves.length;
         const ci = Math.min(wx, hy);
         width = ci * this.columnLeaves.length;
         height = ci * this.rowLeaves.length;
+      } else if (this.layout === 'full') {
+        width = this.columnLeaves.length * (MAX_FONT_SIZE + 2);
+        height = this.rowLeaves.length * (MAX_FONT_SIZE + 2);
       } else if (this.layout === 'squareMatrix') {
         width = Math.min(width, height);
         height = width;
@@ -254,7 +314,7 @@ export default {
       const wx = this.matrixWidth / this.columnLeaves.length - 2;
       const hy = this.matrixHeight / this.rowLeaves.length - 2;
 
-      return Math.min(wx, hy, 12);
+      return Math.min(wx, hy, MAX_FONT_SIZE);
     },
   },
   mounted() {
@@ -273,7 +333,7 @@ export default {
           s.names = [s.name];
         } else {
           s.indices = [].concat(...s.children.map(injectIndices));
-          s.names = [].concat(...s.children.map(c => c.names));
+          s.names = [].concat(...s.children.map((c) => c.names));
           s.name = s.names.join(', ');
         }
         return s.indices;
@@ -282,7 +342,7 @@ export default {
       injectIndices(node);
 
       let root = hierarchy(node,
-        d => (collapsed.has(d) ? [] : (d.children || [])))
+        (d) => (collapsed.has(d) ? [] : (d.children || [])))
         .count()
         .sort((a, b) => b.height - a.height || b.data.index - a.data.index);
 
@@ -301,12 +361,12 @@ export default {
         return null;
       }
       const l = cluster()
-        .size([layoutWidth, layoutHeight * DENDOGRAM_RATIO - this.padding2])
+        .size([layoutWidth, layoutHeight * DENDROGRAM_RATIO - this.padding2])
         .separation(() => 1);
       return l(root);
     },
     updateTree(ref, root, wrapper, config, horizontalLayout) {
-      if (!ref || !config.dendogram || !root) {
+      if (!ref || !config.dendrogram || !root) {
         return;
       }
       const svg = select(ref);
@@ -322,14 +382,14 @@ export default {
       const { hovered, collapsed } = wrapper;
       const { padding } = this;
 
-      edges.classed('selected', d => d.target.data.indices.some(l => hovered.has(l)));
+      edges.classed('selected', (d) => d.target.data.indices.some((l) => hovered.has(l)));
 
-      const renderVerticalLinks = d => `
+      const renderVerticalLinks = (d) => `
         M${d.target.x},${d.target.y + (d.target.children ? 0 : padding)}
         L${d.target.x},${d.source.y}
         L${d.source.x},${d.source.y}
       `;
-      const renderHorizontalLinks = d => `
+      const renderHorizontalLinks = (d) => `
         M${d.target.x + (d.target.children ? 0 : padding)},${d.target.y}
         L${d.source.x},${d.target.y}
         L${d.source.x},${d.source.y}
@@ -337,11 +397,11 @@ export default {
 
       edges.attr('d', horizontalLayout ? renderVerticalLinks : renderHorizontalLinks);
 
-      const innerNodes = root.descendants().filter(d => d.data.indices.length > 1);
+      const innerNodes = root.descendants().filter((d) => d.data.indices.length > 1);
       const inner = svg.select('g.nodes').selectAll('g').data(innerNodes).join((enter) => {
         const r = enter.append('g')
           .html(`<circle r="${padding}"></circle><text><text><title></title>`)
-          .attr('transform', d => `translate(${d.x},${d.y})`);
+          .attr('transform', (d) => `translate(${d.x},${d.y})`);
         r.on('click', (d) => {
           if (wrapper.focus === d.data) {
             wrapper.focus = null;
@@ -368,21 +428,21 @@ export default {
         }
         return collapsed.has(d.data) ? MDI_PLUS_CIRCLE : MDI_MINUS_CIRCLE;
       });
-      inner.select('title').text(d => d.data.name);
-      inner.classed('collapsed', d => collapsed.has(d.data));
-      inner.classed('focused', d => wrapper.focus === d.data);
+      inner.select('title').text((d) => d.data.name);
+      inner.classed('collapsed', (d) => collapsed.has(d.data));
+      inner.classed('focused', (d) => wrapper.focus === d.data);
 
-      inner.attr('transform', d => `translate(${d.x},${d.y})`);
+      inner.attr('transform', (d) => `translate(${d.x},${d.y})`);
     },
     updateColumn() {
-      if (this.columnDendogramHeight === 0) {
+      if (this.columnDendrogramHeight === 0) {
         return;
       }
       this.updateTree(this.$refs.column, this.columnHierarchy, this.column,
         this.columnConfig, true);
     },
     updateRow() {
-      if (this.rowDendogramWidth === 0) {
+      if (this.rowDendrogramWidth === 0) {
         return;
       }
       this.updateTree(this.$refs.row, this.rowHierarchy, this.row, this.rowConfig, false);
@@ -413,12 +473,12 @@ export default {
       });
       const { hovered } = wrapper;
 
-      text.classed('selected', d => d.data.indices.some(l => hovered.has(l)));
+      text.classed('selected', (d) => d.data.indices.some((l) => hovered.has(l)));
 
-      text.select('.label').text(d => d.data.name);
+      text.select('.label').text((d) => d.data.name);
       text.select('.color')
         .classed('hidden', !colorer)
-        .style('background', colorer ? (d => this.combineColor(d.data.names, colorer)) : null);
+        .style('background', colorer ? ((d) => this.combineColor(d.data.names, colorer)) : null);
     },
     updateColumnLabel() {
       this.updateLabel(this.$refs.collabel, this.column, this.columnLeaves,
@@ -451,15 +511,14 @@ export default {
 
       ctx.strokeStyle = 'orange';
 
-
       // work on copy for speed
-      const data = values.map(r => r.slice());
+      const data = values.map((r) => r.slice());
 
-      const cnodeIndices = columns.map(n => n.data.indices);
+      const cnodeIndices = columns.map((n) => n.data.indices);
 
       rows.forEach((rnode, i) => {
         const rnodeIndices = rnode.data.indices;
-        const rowSelected = rnodeIndices.some(s => hoveredRow.has(s));
+        const rowSelected = rnodeIndices.some((s) => hoveredRow.has(s));
         cnodeIndices.forEach((indices, j) => {
           const v = aggregate(data, rnodeIndices, indices, transposed);
           ctx.fillStyle = valueScale(v);
@@ -477,7 +536,7 @@ export default {
       });
 
       cnodeIndices.forEach((indices, j) => {
-        const columnSelected = indices.some(s => hoveredColumn.has(s));
+        const columnSelected = indices.some((s) => hoveredColumn.has(s));
         if (columnSelected) {
           ctx.beginPath();
           ctx.moveTo(j * w, 0);
@@ -498,8 +557,9 @@ export default {
       const w = canvas.width / this.columnLeaves.length;
       const h = canvas.height / this.rowLeaves.length;
 
-      const j = Math.floor(evt.offsetX / w);
-      const i = Math.floor(evt.offsetY / h);
+      const j = Math.floor(Math.max(evt.offsetX, 0) / w);
+      const i = Math.floor(Math.max(evt.offsetY, 0) / h);
+
       const rnode = this.rowLeaves[i].data;
       const cnode = this.columnLeaves[j].data;
       const rowChanged = changeHovered(this.row, rnode.indices);
@@ -516,40 +576,40 @@ export default {
 
     generateImage() {
       const canvas = document.createElement('canvas');
-      const columnDendogram = this.columnConfig.dendogram ? this.height * DENDOGRAM_RATIO : 0;
-      const rowDendogram = this.rowConfig.dendogram ? this.width * DENDOGRAM_RATIO : 0;
+      const columnDendrogram = this.columnConfig.dendrogram ? this.height * DENDROGRAM_RATIO : 0;
+      const rowDendrogram = this.rowConfig.dendrogram ? this.width * DENDROGRAM_RATIO : 0;
 
-      canvas.width = rowDendogram + this.matrixWidth + LABEL_WIDTH;
-      canvas.height = columnDendogram + this.matrixHeight + LABEL_WIDTH;
+      canvas.width = rowDendrogram + this.matrixWidth + LABEL_WIDTH;
+      canvas.height = columnDendrogram + this.matrixHeight + LABEL_WIDTH;
 
       const ctx = canvas.getContext('2d');
       // copy matrix first
       ctx.fillStyle = 'white';
       ctx.fillRect(0, 0, canvas.width, canvas.height);
-      ctx.drawImage(this.$refs.matrix, rowDendogram, columnDendogram);
+      ctx.drawImage(this.$refs.matrix, rowDendrogram, columnDendrogram);
 
       const toWait = [];
 
-      if (columnDendogram > 0) {
-        // column dendogram
+      if (columnDendrogram > 0) {
+        // column dendrogram
         const url = svg2url(this.$refs.column, { font: false, icons: true });
-        const img = new Image(this.width * DENDOGRAM_RATIO, this.matrixHeight);
+        const img = new Image(this.width * DENDROGRAM_RATIO, this.matrixHeight);
         toWait.push(new Promise((resolve) => {
           img.onload = () => {
-            ctx.drawImage(img, rowDendogram, 0);
+            ctx.drawImage(img, rowDendrogram, 0);
             URL.revokeObjectURL(url);
             resolve();
           };
           img.src = url;
         }));
       }
-      if (rowDendogram > 0) {
-        // row dendogram
+      if (rowDendrogram > 0) {
+        // row dendrogram
         const url = svg2url(this.$refs.row, { font: false, icons: true });
-        const img = new Image(this.matrixWidth, this.height * DENDOGRAM_RATIO);
+        const img = new Image(this.matrixWidth, this.height * DENDROGRAM_RATIO);
         toWait.push(new Promise((resolve) => {
           img.onload = () => {
-            ctx.drawImage(img, 0, columnDendogram);
+            ctx.drawImage(img, 0, columnDendrogram);
             URL.revokeObjectURL(url);
             resolve();
           };
@@ -560,7 +620,7 @@ export default {
       // row labels
       {
         ctx.save();
-        ctx.translate(rowDendogram + this.matrixWidth, columnDendogram);
+        ctx.translate(rowDendrogram + this.matrixWidth, columnDendrogram);
         const hi = this.matrixHeight / this.rowLeaves.length;
         const { colorer } = this.rowConfig;
         if (colorer) {
@@ -583,7 +643,7 @@ export default {
       // column labels
       {
         ctx.save();
-        ctx.translate(rowDendogram, columnDendogram + this.matrixHeight);
+        ctx.translate(rowDendrogram, columnDendrogram + this.matrixHeight);
         const wi = this.matrixWidth / this.columnLeaves.length;
         const { colorer } = this.columnConfig;
         if (colorer) {
@@ -613,34 +673,85 @@ export default {
 };
 </script>
 
-<template lang="pug">
-.grid(v-resize:throttle="onResize")
-  svg.column(ref="column", v-show="columnConfig.dendogram",
-      :width="matrixWidth",
-      :height="height * DENDOGRAM_RATIO", xmlns="http://www.w3.org/2000/svg",
-      :data-update="reactiveColumnUpdate")
-    g.edges(:transform="`translate(0,${padding})`")
-    g.nodes(:transform="`translate(0,${padding})`")
-  svg.row(ref="row", v-show="rowConfig.dendogram",
-      :width="width * DENDOGRAM_RATIO",
-      :height="matrixHeight", xmlns="http://www.w3.org/2000/svg",
-      :data-update="reactiveRowUpdate")
-    g.edges(:transform="`translate(${padding},0)`")
-    g.nodes(:transform="`translate(${padding},0)`")
-  canvas.matrix(ref="matrix", :data-update="reactiveMatrixUpdate",
-      @mousemove="canvasMouseMove($event)", @mouseleave="canvasMouseLeave()")
-  .collabel(ref="collabel",
-      :style="{fontSize: fontSize+'px', width: this.matrixWidth+'px', height: LABEL_WIDTH+'px'}",
-      :data-update="reactiveColumnLabelUpdate")
-  .rowlabel(ref="rowlabel",
-      :style="{fontSize: fontSize+'px', width: LABEL_WIDTH+'px', height: this.matrixHeight+'px'}",
-      :data-update="reactiveRowLabelUpdate")
-  .legend-wrapper(v-show="columnConfig.dendogram && rowConfig.dendogram")
-    .legend(:data-from="legendDomain[0]", :data-to="legendDomain[1]",
-        :style="{background: legendGradient}")
+<template>
+  <div
+    v-resize:throttle="onResize"
+    :class="{
+      'grid': true,
+      'grid--large': layout === 'full',
+    }"
+  >
+    <svg
+      v-show="columnConfig.dendrogram"
+      ref="column"
+      class="column"
+      :width="matrixWidth"
+      :height="height * DENDROGRAM_RATIO"
+      xmlns="http://www.w3.org/2000/svg"
+      :data-update="reactiveColumnUpdate"
+    >
+      <g
+        class="edges"
+        :transform="`translate(0,${padding})`"
+      />
+      <g
+        class="nodes"
+        :transform="`translate(0,${padding})`"
+      />
+    </svg>
+    <svg
+      v-show="rowConfig.dendrogram"
+      ref="row"
+      class="row"
+      :width="width * DENDROGRAM_RATIO"
+      :height="matrixHeight"
+      xmlns="http://www.w3.org/2000/svg"
+      :data-update="reactiveRowUpdate"
+    >
+      <g
+        class="edges"
+        :transform="`translate(${padding},0)`"
+      />
+      <g
+        class="nodes"
+        :transform="`translate(${padding},0)`"
+      />
+    </svg>
+    <canvas
+      ref="matrix"
+      class="matrix"
+      :data-update="reactiveMatrixUpdate"
+      @mousemove="canvasMouseMove($event)"
+      @mouseleave="canvasMouseLeave()"
+    />
+    <div
+      ref="collabel"
+      class="collabel"
+      :style="{fontSize: fontSize+'px', width: matrixWidth+'px', height: LABEL_WIDTH+'px'}"
+      :data-update="reactiveColumnLabelUpdate"
+    />
+    <div
+      ref="rowlabel"
+      class="rowlabel"
+      :style="{fontSize: fontSize+'px', width: LABEL_WIDTH+'px', height: matrixHeight+'px'}"
+      :data-update="reactiveRowLabelUpdate"
+    />
+    <div
+      v-show="columnConfig.dendrogram && rowConfig.dendrogram"
+      class="legend-wrapper"
+    >
+      <div
+        class="legend"
+        :data-from="legendDomain[0]"
+        :data-to="legendDomain[1]"
+        :style="{background: legendGradient}"
+      />
+    </div>
+  </div>
 </template>
 
-<style scoped>
+<style scoped lang="scss">
+
 .grid {
   position: absolute;
   top: 4px;
@@ -648,19 +759,24 @@ export default {
   right: 8px;
   bottom: 8px;
   display: grid;
-  grid-template-areas: "legend column dl"
-    "row matrix rlabel"
-    "rc clabel ll";
+  grid-template-areas: "legend column dl" "row matrix rlabel" "rc clabel ll";
   justify-content: center;
   align-content: center;
+
+  &--large {
+    position: relative;
+    justify-content: left;
+  }
 }
 
 .column {
   grid-area: column;
 }
+
 .row {
   grid-area: row;
 }
+
 .matrix {
   grid-area: matrix;
 }
@@ -675,21 +791,18 @@ export default {
   height: 1.5em;
   margin: 2em;
   position: relative;
+  &::before {
+    content: attr(data-from);
+    position: absolute;
+    top: 100%;
+  }
+  &::after {
+    content: attr(data-to);
+    position: absolute;
+    top: 100%;
+    right: 0;
+  }
 }
-
-.legend::before {
-  content: attr(data-from);
-  position: absolute;
-  top: 100%;
-}
-
-.legend::after {
-  content: attr(data-to);
-  position: absolute;
-  top: 100%;
-  right: 0;
-}
-
 
 .collabel {
   grid-area: clabel;
@@ -697,22 +810,29 @@ export default {
   justify-content: center;
   overflow: hidden;
   line-height: normal;
-}
-
-.collabel >>> div {
-  flex-direction: column;
-}
-
-.collabel >>> .label {
-  text-align: right;
-  writing-mode: tb;
-  transform: rotate(-180deg);
-}
-
-.collabel >>> .color {
-  align-self: stretch;
-  height: 5px;
-  margin-bottom: 2px;
+  /deep/ div {
+    flex-direction: column;
+    flex: 1 1 0;
+    display: flex;
+    align-items: center;
+  }
+  /deep/ .label {
+    text-align: right;
+    writing-mode: tb;
+    transform: rotate(-180deg);
+    flex: 1 1 0;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+  /deep/ .color {
+    align-self: stretch;
+    height: 5px;
+    margin-bottom: 2px;
+  }
+  /deep/ .selected {
+    color: orange;
+  }
 }
 
 .rowlabel {
@@ -721,78 +841,63 @@ export default {
   flex-direction: column;
   justify-content: center;
   line-height: normal;
+  /deep/ .color {
+    align-self: stretch;
+    width: 5px;
+    min-width: 5px;
+    margin-right: 2px;
+  }
+  /deep/ .color.hidden {
+    display: none;
+  }
+  /deep/ div {
+    flex: 1 1 0;
+    display: flex;
+    align-items: center;
+  }
+  /deep/ .label {
+    flex: 1 1 0;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+  /deep/ .selected {
+    color: orange;
+  }
 }
 
-
-.rowlabel >>> .color {
-  align-self: stretch;
-  width: 5px;
-  min-width: 5px;
-  margin-right: 2px;
+.edges {
+  /deep/ path {
+    fill: none;
+    stroke-width: 1;
+    stroke: black;
+  }
+  /deep/ path.selected {
+    stroke: orange;
+  }
 }
 
-.rowlabel >>> .color.hidden {
-  display: none;
+.nodes {
+  /deep/ g {
+    opacity: 0;
+    font: normal normal normal 24px/1 "Material Design Icons";
+    fill: black;
+    cursor: pointer;
+    font-size: 150%;
+    user-select: none;
+    text-anchor: middle;
+    dominant-baseline: central;
+    /deep/ circle {
+      fill: white;
+    }
+    &:hover {
+      opacity: 1;
+      fill: orange;
+      /deep/ text {
+        fill: orange;
+      }
+    }
+  }
 }
-
-.collabel >>> div,
-.rowlabel >>> div {
-  flex: 1 1 0;
-  display: flex;
-  align-items: center;
-}
-
-.collabel >>> .label,
-.rowlabel >>> .label {
-  flex: 1 1 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.rowlabel >>> .selected,
-.collabel >>> .selected {
-  color: orange;
-}
-
-.edges >>> path {
-  fill: none;
-  stroke-width: 1;
-  stroke: black;
-}
-
-.edges >>> path.selected {
-  stroke: orange;
-}
-
-.nodes >>> g {
-  opacity: 0;
-  font: normal normal normal 24px/1 "Material Design Icons";
-  fill: black;
-  cursor: pointer;
-  font-size: 150%;
-  user-select: none;
-  text-anchor: middle;
-  dominant-baseline: central;
-}
-
-.nodes >>> g > circle {
-  fill: white;
-}
-
-.nodes >>> g:hover {
-  opacity: 1;
-  fill: orange;
-}
-
-.nodes >>> g:hover > text {
-  fill: orange;
-}
-
-.nodes >>> .collapsed,
-.nodes >>> .focused {
-  opacity: 1;
-}
-
 
 </style>
