@@ -1,11 +1,13 @@
 <script>
 import LoadingsPlot from './LoadingsPlot.vue';
+import LoadingsPlotHelp from './help/LoadingsPlotHelp.vue';
 import VisTile from './VisTile.vue';
 import plotData from './mixins/plotData';
 
 export default {
   components: {
     LoadingsPlot,
+    LoadingsPlotHelp,
     VisTile,
   },
 
@@ -39,36 +41,70 @@ export default {
 
 </script>
 
-<template lang="pug">
-vis-tile(v-if="plot", title="PCA Loadings Plot", :loading="plot.loading", svg-download)
-  loadings-plot(
-      :points="plot.data",
-      :pc-x="pcX",
-      :pc-y="pcY",
-      :show-crosshairs="showCrosshairs")
-  template(v-slot:controls)
-    v-menu(bottom, offset-y, left, :min-width="150", :close-on-content-click="false")
-      template(v-slot:activator="{ on }")
-        v-btn(v-on="on", icon)
-          v-icon.mdi.mdi-dots-vertical
-
-      v-card.pa-1(flat)
-        v-layout(column)
-          v-text-field.pa-2(
-              hide-details,
-              type="number",
-              label="PC (X Axis)",
-              min="1",
-              outline,
-              v-model="pcXval")
-          v-text-field.pa-2(
-              hide-details,
-              type="number",
-              label="PC (Y Axis)",
-              min="1",
-              outline,
-              v-model="pcYval")
-          v-switch.ma-2(hide-details, v-model="showCrosshairs", label="Show crosshairs")
-  template(v-slot:help)
-    include help/LoadingsPlotHelp.pug
+<template>
+  <vis-tile
+    v-if="plot"
+    title="PCA Loadings Plot"
+    :loading="plot.loading"
+    svg-download
+  >
+    <loadings-plot
+      :points="plot.data"
+      :pc-x="pcX"
+      :pc-y="pcY"
+      :show-crosshairs="showCrosshairs"
+    />
+    <template v-slot:controls>
+      <v-menu
+        bottom
+        offset-y
+        left
+        :min-width="150"
+        :close-on-content-click="false"
+      >
+        <template v-slot:activator="{ on }">
+          <v-btn
+            icon
+            v-on="on"
+          >
+            <v-icon class="mdi mdi-dots-vertical" />
+          </v-btn>
+        </template>
+        <v-card
+          class="pa-1"
+          flat
+        >
+          <v-layout column>
+            <v-text-field
+              v-model="pcXval"
+              class="pa-2"
+              hide-details
+              type="number"
+              label="PC (X Axis)"
+              min="1"
+              outline
+            />
+            <v-text-field
+              v-model="pcYval"
+              class="pa-2"
+              hide-details
+              type="number"
+              label="PC (Y Axis)"
+              min="1"
+              outline
+            />
+            <v-switch
+              v-model="showCrosshairs"
+              class="ma-2"
+              hide-details
+              label="Show crosshairs"
+            />
+          </v-layout>
+        </v-card>
+      </v-menu>
+    </template>
+    <template v-slot:help>
+      <LoadingsPlotHelp />
+    </template>
+  </vis-tile>
 </template>
