@@ -1,66 +1,71 @@
-<script>
-import { format } from 'd3-format';
+<script lang="ts">
+import {
+  PropType, defineComponent, computed, reactive, ref,
+} from '@vue/composition-api';
+import { format as d3Format } from 'd3-format';
 import { colors } from '../../utils/constants';
 import { textColor } from '../../utils';
 
-export default {
+export interface Item {
+  Metabolite: string;
+  color: string;
+  [key: string]: number | string;
+}
 
+interface Header {
+  text: string;
+  align?: string;
+  value: string;
+  isLabel: boolean;
+}
+
+export default defineComponent({
   props: {
     items: {
-      type: Array,
+      type: Array as PropType<Item[]>,
       required: true,
     },
     value: {
-      type: Array,
+      type: Array as PropType<string[]>,
       required: true,
     },
     headers: {
-      type: Array,
+      type: Array as PropType<Header[]>,
       required: true,
     },
     threshold: {
       type: Number,
-      required: false,
       default: 0.05,
     },
+    noDataAvailableMsg: { // allows displaying a custom message when table has no data.
+      type: String,
+      default: 'No data available',
+    },
   },
 
-  data() {
-    return {
-      format: format('.4e'),
-      pagination: {
-        rowsPerPage: -1,
-      },
-      // need to use this workaround, since `header-cell` as slot name doesn't work
-      // and `headerCell` isn't allowed by the linter
-      headercell: 'headerCell',
-      colors,
-    };
-  },
-
-  computed: {
-    selectedItems: {
+  setup(props, { emit }) {
+    const format = ref(d3Format('.4e'));
+    const pagination = reactive({ rowsPerPage: -1 });
+    const selectedItems = computed<Item[]>({
       get() {
-        const s = new Set(this.value);
-        return this.items.filter(d => s.has(d.Metabolite));
+        const s = new Set(props.value);
+        return props.items.filter((d) => s.has(d.Metabolite));
       },
       set(newValue) {
-        const s = newValue.map(d => d.Metabolite);
-        this.$emit('input', s);
+        const s = newValue.map((d) => d.Metabolite);
+        emit('input', s);
       },
-    },
-  },
+    });
 
-  methods: {
-    isInteresting(value) {
-      return value < this.threshold;
-    },
-    toggleHighlighted(header, evt, add) {
+    function isInteresting(value: number) {
+      return value < props.threshold;
+    }
+    function toggleHighlighted(header: Header, evt: Event, add: boolean) {
       evt.preventDefault();
       evt.stopPropagation();
 
-      const highlighted = this.items.filter(item => this.isInteresting(item[header.value]));
-      const current = this.selectedItems.slice();
+      const highlighted = props.items.filter((item) => isInteresting(item[header.value] as number));
+      const current = selectedItems.value.slice();
       const currentLookup = new Set(current);
       highlighted.forEach((item) => {
         if (add && !currentLookup.has(item)) {
@@ -70,9 +75,9 @@ export default {
           current.splice(index, 1);
         }
       });
-      this.selectedItems = current;
-    },
-    cellStyle(item) {
+      selectedItems.value = current;
+    }
+    function cellStyle(item: Item) {
       if (!item.color) {
         return null;
       }
@@ -80,33 +85,86 @@ export default {
         backgroundColor: item.color,
         color: textColor(item.color),
       };
-    },
+    }
+    return {
+      format,
+      pagination,
+      colors,
+      selectedItems,
+      isInteresting,
+      toggleHighlighted,
+      cellStyle,
+    };
   },
-};
+});
 </script>
 
-<template lang="pug">
-v-data-table.elevation-1.main(:headers="headers", :items="items", disable-initial-sort,
-    item-key="Metabolite", :pagination.sync="pagination",
-    v-model="selectedItems", select-all)
-  template(v-slot:[headercell]="{header}")
-    | {{header.text}}
-    v-btn.toggle(icon, small, @click="toggleHighlighted(header, $event, true)",
-        title="Adds the highlighted Metabolites to the selected set",
-        v-if="!header.isLabel")
-      span.mdi(:class="{ [$vuetify.icons.plusMultiple]: true }")
-    v-btn.toggle(icon, small, @click="toggleHighlighted(header, $event, false)",
-        title="Removes the highlighted Metabolites from the selected set",
-        v-if="!header.isLabel")
-      span.mdi(:class="{ [$vuetify.icons.minusMultiple]: true }")
-
-  template(#items="props")
-    td.cell
-      v-checkbox(v-model="props.selected", hide-details, :color="colors.selected")
-    td.cell(:style="cellStyle(props.item)") {{ props.item.Metabolite }}
-    td.cell.text-xs-right(v-for="c in headers.slice(1)",
-        :class="isInteresting(props.item[c.value]) ? 'highlight' : ''")
-      | {{ format(props.item[c.value]) }}
+<template>
+  <v-data-table
+    v-model="selectedItems"
+    class="elevation-1 main"
+    :headers="headers"
+    :items="items"
+    disable-initial-sort
+    item-key="Metabolite"
+    :pagination.sync="pagination"
+    select-all
+  >
+    <template v-slot:headerCell="{ header }">
+      {{ header.text }}<v-btn
+        v-if="!header.isLabel"
+        class="toggle"
+        icon
+        small
+        title="Adds the highlighted Metabolites to the selected set"
+        @click="toggleHighlighted(header, $event, true)"
+      >
+        <span
+          class="mdi"
+          :class="{ [$vuetify.icons.plusMultiple]: true }"
+        />
+      </v-btn>
+      <v-btn
+        v-if="!header.isLabel"
+        class="toggle"
+        icon
+        small
+        title="Removes the highlighted Metabolites from the selected set"
+        @click="toggleHighlighted(header, $event, false)"
+      >
+        <span
+          class="mdi"
+          :class="{ [$vuetify.icons.minusMultiple]: true }"
+        />
+      </v-btn>
+    </template>
+    <template #items="props">
+      <td class="cell">
+        <v-checkbox
+          v-model="props.selected"
+          hide-details
+          :color="colors.selected"
+        />
+      </td>
+      <td
+        class="cell"
+        :style="cellStyle(props.item)"
+      >
+        {{ props.item.Metabolite }}
+      </td>
+      <td
+        v-for="c in headers.slice(1)"
+        :key="c.value"
+        class="cell text-xs-right"
+        :class="isInteresting(props.item[c.value]) ? 'highlight' : ''"
+      >
+        {{ format(props.item[c.value]) }}
+      </td>
+    </template>
+    <template slot="no-data">
+      {{ noDataAvailableMsg }}
+    </template>
+  </v-data-table>
 </template>
 
 <style scoped>
@@ -139,5 +197,4 @@ v-data-table.elevation-1.main(:headers="headers", :items="items", disable-initia
 .main >>> thead > tr > th:hover .toggle {
   opacity: 0.6;
 }
-
 </style>
