@@ -1,92 +1,63 @@
-<script>
+<script lang="ts">
+import { defineComponent, toRef } from '@vue/composition-api';
 import ScorePlot from '@/components/vis/ScorePlot.vue';
+import ScorePlotHelp from '@/components/vis/help/ScorePlotHelp.vue';
 import VisTile from '@/components/vis/VisTile.vue';
-import plotData from '@/components/vis/mixins/plotData';
+import usePlotData from '../use/usePlotData';
 
-export default {
-  components: {
-    ScorePlot,
-    VisTile,
-  },
-
-  mixins: [
-    plotData('pca'),
-  ],
-
+export default defineComponent({
   props: {
     id: {
       type: String,
       required: true,
     },
-
     pcX: {
       type: Number,
-      validator: prop => Number.isInteger(prop),
       required: true,
     },
-
     pcY: {
       type: Number,
-      validator: prop => Number.isInteger(prop),
       required: true,
     },
-
     showEllipses: {
       type: Boolean,
       required: true,
     },
   },
-
-  computed: {
-    ready() {
-      return this.$store.getters.ready(this.id);
-    },
-
-    pcCoords() {
-      return this.maybeData('x', []);
-    },
-
-    rowLabels() {
-      return this.maybeData('rows', []);
-    },
-
-    groupLabels() {
-      return this.maybeData('labels', {});
-    },
-
-    eigenvalues() {
-      return this.maybeData('sdev', []);
-    },
-
-    columns() {
-      return this.dataset.column.data;
-    },
+  components: {
+    ScorePlot,
+    ScorePlotHelp,
+    VisTile,
   },
-
-  methods: {
-    maybeData(key, dflt) {
-      const {
-        plot,
-      } = this;
-
-      return plot.data ? plot.data[key] : dflt;
-    },
+  setup(props) {
+    const { plot, dataset } = usePlotData(toRef(props, 'id'), 'pca');
+    return {
+      plot,
+      dataset,
+    };
   },
-};
+});
 </script>
 
-<template lang="pug">
-vis-tile(title="PCA Score Plot", :loading="plot.loading", svg-download)
-  score-plot(
-      :pc-coords="pcCoords",
-      :row-labels="rowLabels",
-      :colors="dataset.groupLevels",
-      :group-labels="groupLabels",
-      :eigenvalues="eigenvalues",
-      :columns="columns",
-      :pc-x="pcX",
-      :pc-y="pcY",
-      :show-ellipses="showEllipses")
-  template(v-slot:help)
-    include ../help/ScorePlotHelp.pug
+<template>
+  <vis-tile
+    title="PCA Score Plot"
+    :loading="plot.loading"
+    svg-download
+  >
+    <score-plot
+      v-if="plot.data && dataset"
+      :pc-coords="plot.data.x"
+      :row-labels="plot.data.rows"
+      :colors="dataset.groupLevels"
+      :group-labels="plot.data.labels"
+      :eigenvalues="plot.data.sdev"
+      :columns="dataset.column.data"
+      :pc-x="pcX"
+      :pc-y="pcY"
+      :show-ellipses="showEllipses"
+    /><template v-slot:help>
+      <ScorePlotHelp />
+    </template>
+  </vis-tile>
 </template>
