@@ -1,60 +1,100 @@
-<script>
+<script lang="ts">
+import {
+  defineComponent, computed, ref, toRef,
+} from '@vue/composition-api';
 import ScreePlot from '@/components/vis/ScreePlot.vue';
 import VisTile from '@/components/vis/VisTile.vue';
-import plotData from './mixins/plotData';
+import ScreePlotHelp from './help/ScreePlotHelp.vue';
+import usePlotData from './use/usePlotData';
 
-export default {
+export default defineComponent({
+  props: {
+    id: {
+      type: String,
+      required: true,
+    },
+  },
+
   components: {
     ScreePlot,
+    ScreePlotHelp,
     VisTile,
   },
 
-  mixins: [plotData('pca')],
+  setup(props) {
+    const showCutoffs = ref(true);
+    const numComponentsText = ref('10');
+    const numComponents = computed(() => Number.parseInt(numComponentsText.value, 10));
 
-  props: {
-    id: {
-      required: true,
-      type: String,
-    },
-  },
+    const { plot } = usePlotData(toRef(props, 'id'), 'pca');
 
-  data() {
     return {
-      numComponentsText: '10',
-      showCutoffs: true,
+      numComponentsText,
+      numComponents,
+      showCutoffs,
+      plot,
     };
   },
-
-  computed: {
-    numComponents() {
-      return Number.parseInt(this.numComponentsText, 10);
-    },
-  },
-};
+});
 </script>
 
-<template lang="pug">
-vis-tile(v-if="plot", title="PCA Scree Plot", :loading="plot.loading", svg-download)
-  scree-plot(
-      :eigenvalues="getPlotDataProperty('sdev')",
-      :num-components="numComponents",
-      :show-cutoffs="showCutoffs")
-  template(v-slot:controls)
-    v-menu(bottom, offset-y, left, :min-width="150", :close-on-content-click="false")
-      template(v-slot:activator="{ on }")
-        v-btn(v-on="on", icon)
-          v-icon.mdi.mdi-dots-vertical
-
-      v-card.pa-1(flat)
-        v-layout.px-2(column)
-          v-text-field.py-2(
-              hide-details,
-              type="number",
-              label="Principal Components",
-              min="1",
-              outline,
-              v-model="numComponentsText")
-          v-switch.py-2(v-model="showCutoffs", label="Diagnostic cutoffs", hide-details)
-  template(v-slot:help)
-    include help/ScreePlotHelp.pug
+<template>
+  <vis-tile
+    v-if="plot"
+    title="PCA Scree Plot"
+    :loading="plot.loading"
+    svg-download
+  >
+    <scree-plot
+      v-if="plot.data"
+      :eigenvalues="plot.data.sdev"
+      :num-components="numComponents"
+      :show-cutoffs="showCutoffs"
+    /><template v-slot:controls>
+      <v-menu
+        bottom
+        offset-y
+        left
+        :min-width="150"
+        :close-on-content-click="false"
+      >
+        <template v-slot:activator="{ on }">
+          <v-btn
+            icon
+            v-on="on"
+          >
+            <v-icon class="mdi mdi-dots-vertical" />
+          </v-btn>
+        </template>
+        <v-card
+          class="pa-1"
+          flat
+        >
+          <v-layout
+            class="px-2"
+            column
+          >
+            <v-text-field
+              v-model="numComponentsText"
+              class="py-2"
+              hide-details
+              type="number"
+              label="Principal Components"
+              min="1"
+              outline
+            />
+            <v-switch
+              v-model="showCutoffs"
+              class="py-2"
+              label="Diagnostic cutoffs"
+              hide-details
+            />
+          </v-layout>
+        </v-card>
+      </v-menu>
+    </template>
+    <template v-slot:help>
+      <ScreePlotHelp />
+    </template>
+  </vis-tile>
 </template>
