@@ -50,11 +50,11 @@ export default {
 
       if (this.metaboliteFilter) {
         const filter = this.metaboliteFilter.apply;
-        data = data.filter(row => filter(row.Metabolite));
+        data = data.filter((row) => filter(row.Metabolite));
       }
       if (this.metaboliteColor) {
         const colorer = this.metaboliteColor.apply;
-        data = data.map(row => ({ ...row, color: colorer(row.Metabolite) }));
+        data = data.map((row) => ({ ...row, color: colorer(row.Metabolite) }));
       }
 
       return {
@@ -72,7 +72,7 @@ export default {
 </script>
 
 <template lang="pug">
-vis-tile-large(title="Wilcoxon Test", :loading="plot.loading", expanded)
+vis-tile-large(title="Wilcoxon Test", analysis-path="wilcoxon", :loading="plot.loading", expanded)
   template(#controls)
     metabolite-filter(:dataset="dataset", v-model="metaboliteFilter", hide-selection)
     metabolite-colorer(:dataset="dataset", v-model="metaboliteColor", hide-selection,
@@ -88,7 +88,12 @@ vis-tile-large(title="Wilcoxon Test", :loading="plot.loading", expanded)
     v-btn(flat, dark, block, @click="downloadTable")
       v-icon.mr-2 {{ $vuetify.icons.save }}
       | Download Table
-  wilcoxon-plot(v-if="plot.data", :data="tableData", :threshold="threshold", v-model="selected")
+  wilcoxon-plot(
+      v-if="plot.data",
+      :data="tableData",
+      :threshold="threshold",
+      :error-msg="plot.data.error || ''",
+      v-model="selected")
 </template>
 
 <style scoped>
